fixed ratio, moved some functions around
